fix: template/DeepLinkBooks
